Add vlan support.
With the move to generic hardware and away from Broadcom swtiches it was necessary to make some backend changes.
Every nic is now independent and each individual one needs to be added to their respective bridges.
The WebUI should remain functionally the same, transparent to the end user.
We are currently still limited to 5 network ports in the WebUI, although this could be increased in the future.
It should however be possible to manage extra nics with some nvram manipulation.

The default port configuration is as follows.
Wan0 -> eth0
br0 - > eth1 eth2 eth3 eth4