Status-Overview.asp: Add more stats and fix Ethernet Port State.
Running on bare metal the Ethernet Port State should correctly reflect the Link state and Speed.

In a VM it will depend on how you set it up. If your system supports vt-d/IOMMU/pci passthrough,
you can passthrough the nic directly to the VM and you will see the Correct Link State and Speed.
If however you assign virtual nics that are bridged to the real nics, those nics will appear
to always be up, you will see "AUTO" in place of a link speed. In that scenario you will have to
query the host machine to see the actual link states and speeds.