Default debug to console to off. Seems to be a bit buggy sometimes.
This is the tomato built-in debugging functionality.
It opens a pipe to the console, but it seems to causes issues on occasion.
I'm guessing something with parallel writes going on, but idk for sure.
I'll probably re-implement my own debug to console which didn't seem to have issues with the console.