/sbin/upgrade: This commit removes obsoleted files when performing an upgrade
Up to now when performing upgrades we have just been extracting the new filesystem over the old one.
If a file was removed from the image, it was not deleted from the filesystem.
This commit now deletes files that have been eliminiated from the system image.

The current directories that are monitored are
/bin /sbin /lib /libexec /usr/bin /usr/sbin /usr/lib /usr/libexec /usr/share /boot /rom /www

Users should abstain from adding any customization to these directories because it will be lost in an upgrade
Instead they should use those directories that have traditionally been present and read/write/mountable in Tomato for users to use. These are
/opt, /mmc, /nas, /jffs, cifs1, cifs2, /bkp, /media, /tftpboot